+++ gframe/single_mode.h
@@ -33,9 +33,9 @@ public:
 	static int SinglePlayThread(DuelOptions duelOptions);
 	static bool SinglePlayAnalyze(CoreUtils::Packet packet);
 	
-	static void SinglePlayRefresh(int player, int location, int flag = 0x2f81fff);
-	static void SinglePlayRefresh(int flag = 0x2f81fff);
-	static void SinglePlayRefreshSingle(int player, int location, int sequence, int flag = 0x2f81fff);
+	static void SinglePlayRefresh(uint8_t player, uint8_t location, uint32_t flag = 0x2f81fff);
+	static void SinglePlayRefresh(uint32_t flag = 0x2f81fff);
+	static void SinglePlayRefreshSingle(uint8_t player, uint8_t location, uint32_t sequence, uint32_t flag = 0x2f81fff);
 	static void SinglePlayReload();
 	static Signal singleSignal;
